requirements.txt
# PDF处理库
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0

# 基础依赖
numpy>=1.21.0
pandas>=1.3.0

# HTTP请求库（用于调用embedding API）
requests>=2.28.0

# 配置管理
python-dotenv>=0.19.0

# 日志和工具
tqdm>=4.64.0

# 文本处理
jieba>=0.42.1

# 向量存储
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.10.1
chromadb>=0.4.0

# Web框架（如果需要API服务）
fastapi>=0.100.0
uvicorn>=0.23.0

# JSON处理
orjson>=3.8.0